Add the WPS version number to metadata of files created by cams2wrfchem.py (#67) (#81)
In WRF (share/input_wrf.F), when reading the `wrfchemi` files, the model checks that they were created for WRFV4 by reading the TITLE global attribute and searching for the substring " V4.". WRF stops if this substring is not found.

This commit makes the `cams2wrfchem.py` script add this information ("created ... for WRF Vxxx" ) in the TITLE attribute of the files it creates, where the version number is read from the `wrfinput` files used as input to this script.