Design a test case and create infrastructure to run it (#93)
It is important for us to have a "fixed" test case that we can use to
systematically test new WRF-Chem-Polar developments. We decided that
this test case would also become our default test case in WRF-infra.

The design of this test case has been agreed upon by the team during our
2025 November hackathon. The specifics of this test case are detailed in
issue #83.

---------

Co-authored-by: ruthprice <22415989+ruthprice@users.noreply.github.com>